% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqercc.R
\name{qqercc}
\alias{qqercc}
\title{qqercc}
\usage{
qqercc(
  ercc_file,
  control_group,
  comparisons = NULL,
  method = "mean",
  format = "bar",
  color = NULL
)
}
\arguments{
\item{ercc_file}{raw counts}

\item{control_group}{control group}

\item{comparisons}{t.test}

\item{method}{mean,median,deseq2}

\item{format}{bar,box,dot}

\item{color}{diy}
}
\value{
ercc_normalized file,relative_to_control_picture
}
\description{
qqercc
}
\examples{
result <- qqercc("ercc.xlsx", control_group = "control", method = "deseq2")
}
